services/auth_service/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
sqlalchemy
asyncpg
passlib[bcrypt]
python-jose[cryptography]
pydantic[email]
aiosmtplib
python-multipart
redis[asyncio]
bcrypt==4.0.1
celery[redis]==5.3.6
httpx
